이미지, 썸네일 파일 삭제 기능 완료
1. goodsEnroll.jsp jQuery.data 활용하여 파일 삭제 기능 추가 완료

2. FileManager 이미지 파일 삭제 메서드 오류 해결
  - 기존 파일삭제 기능 실행시 파일, 폴더 존재 여부 확인하지 않고 삭제
    file.isDirectory 메서드 활용하여 검사를 수행 후 삭제하도록 수정
    해당 메서드는 경로가 폴더인지 확인하는 메서드이므로 false값 반환
    exists() 메서드[경로, 폴더 존재여부 확인]를 사용하여 오류 해결

+++ mvcPractice04/src/main/java/com/spring/shop/util/FileManager.java
@@ -180,17 +180,24 @@ public class FileManager {
 			
 			// 원본 파일 경로
 			String originFilePath = file.getAbsolutePath().replace("t_", "");
-						
-			// 썸네일 삭제 
-			file.delete();
-						
-			// 원본 파일 객체 생성
-			file = new File(originFilePath);
-						
-			// 원본 파일 삭제
-			file.delete();
+
+			if(file.exists()) {
+				
+				// 썸네일 삭제
+				file.delete();
+				
+				// 원본 파일 객체 생성
+				file = new File(originFilePath);
+				
+				// 원본 파일 삭제
+				file.delete();
+
+				return true;
+			}
 			
-			return true;
+			// 파일 또는 디렉토리가 존재하지 않을 경우
+			return false;		
+			
 		} catch(Exception e) {
 			e.printStackTrace();
 			return false;
